Move requireAuth to user
